add Timecode.div
